Change handling of c++ calling a director that throws a python exception. We want to use the original exception, rather than swallowing it up into a new exception.

+++ lib/swig_to_python.h
@@ -128,7 +128,7 @@ template<typename T> inline PyObject*
 swig_to_python(const std::vector<boost::shared_ptr<T> >& V)
 {
   PyObject* res = PyList_New(V.size());
-  for(int i = 0; i < V.size(); ++i)
+  for(int i = 0; i < (int) V.size(); ++i)
     PyList_SetItem(res, i, swig_to_python(V[i]));
   return res;
 }
